// src/interfaces/RickAndMortyInterace.tsx
// Generated by https://quicktype.io

export interface RickAndMortyReponse {
  info: Info;
  results: Characters[];
}

export interface Info {
  count: number;
  pages: number;
  next: string;
  prev: null;
}

export interface Characters {
  id: number;
  name: string;
  species: Species;
  origin: Location;
  location: Location;
  image: string;
  status: Status;
  episode: string[];
}

export enum Gender {
  Female = 'Female',
  Male = 'Male',
  Unknown = 'unknown',
}

export interface Location {
  name: string;
  url: string;
}

export enum Species {
  Alien = 'Alien',
  Human = 'Human',
}

export enum Status {
  Alive = 'Alive',
  Dead = 'Dead',
  Unknown = 'unknown',
}

// Generated by https://quicktype.io

export interface LocationResponse {
  info: InfoLocation;
  results: LocationResp[];
}

export interface InfoLocation {
  count: number;
  pages: number;
  next: string;
  prev: null;
}

export interface LocationResp {
  id: number;
  name: string;
  type: string;
  dimension: string;
  residents: string[];
  url: string;
  created: string;
}

// Generated by https://quicktype.io

export interface EpisodeResponse {
  info: InfoEpisode;
  results: ResultEpisode[];
}

export interface InfoEpisode {
  count: number;
  pages: number;
  next: string;
  prev: null;
}

export interface ResultEpisode {
  id: number;
  name: string;
  air_date: string;
  episode: string;
  characters: string[];
  url: string;
  created: string;
}

// Generated by https://quicktype.io

export interface ResultInfoEpisode {
  id: number;
  name: string;
  air_date: string;
  episode: string;
  characters: string[];
  url: string;
  created: string;
}

export interface CharactersImage {
  id: number;
  image: string;
  name: string;
}